fix integration

+++ app/src/main/java/com/payrexx/taptopayecr/dto/RefundDataDTO.java
@@ -3,5 +3,5 @@ package com.payrexx.taptopayecr.dto;
 public class RefundDataDTO implements DataDto {
     public String transaction_id = "";
     public String extras = "";
-    public int amount = 0;
+    public float amount = 0f;
 }
